Ticket #3215 - Incoroprated code review comments
